requirements.txt
altair==6.0.0
beautifulsoup4==4.14.3
dash==4.0.0
dash-bootstrap-components==2.0.4
kroger-api==0.2.0
matplotlib==3.10.8
numpy==2.2.6
pandas==2.3.3
plotly==6.6.0
pyarrow==23.0.1
python-dotenv==1.2.2
requests==2.32.5
streamlit==1.55.0
tqdm==4.67.3

# notebooks
jupyterlab
ipykernel

# development
pytest==9.0.2
hypothesis==6.151.9
pylint==4.0.4
pipreqs==0.5.0
build==1.4.0
coverage==7.6.1